Verify source version before parsing PR
Non-PR hooks can trigger builds in which the CODEBUILD_SOURCE_VERSION
would be populated as the commit SHA. Our current logic dictates that
that commit SHA would be passed onto CodeCov as the PR number, which
would always be an invalid value (must be a number, true, or falsy).

This updates the logic so that it will ensure that the environment
variable being checked actually matches the "pr/" pattern before trying
to use it. If it does not, it will return undefined.